curriesportal/requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
blivet==3.4.0
Brlapi==0.6.7
certifi==2022.9.24
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.0.0
chrome-gnome-shell==0.0.0
contextvars==2.4
cupshelpers==1.0
dasbus==1.2
dbus-python==1.2.4
decorator==4.2.1
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.6
Django==3.2.16
django-import-export==2.7.1
django-jazzmin==2.6.0
django-userforeignkey==0.5.0
et-xmlfile==1.1.0
file-magic==0.3.0
filelock==3.4.1
gpg==1.13.1
html5lib==0.999999999
idna==3.4
immutables==0.19
importlib-metadata==4.8.3
importlib-resources==5.4.0
initial-setup==0.3.81.7
isc==2.0
langtable==0.0.51
libcomps==0.1.18
lxml==4.2.3
MarkupPy==1.14
nftables==0.1
ntplib==0.3.3
odfpy==1.4.1
openpyxl==3.0.10
ordered-set==2.0.2
perf==0.1
pexpect==4.3.1
pid==2.1.1
Pillow==8.4.0
platformdirs==2.4.0
ply==3.9
productmd==1.11
psutil==5.4.3
ptyprocess==0.5.2
pwquality==1.4.4
pycairo==1.16.3
pycparser==2.21
pycups==1.9.72
pycurl==7.43.0.2
pydbus==0.6.0
pyenchant==2.0.0
PyGObject==3.28.3
pykickstart==3.16.15
pyparted==3.11.7
PySocks==1.6.8
python-dateutil==2.6.1
python-linux-procfs==0.7.0
python-meh==0.47.2
pytz==2022.2.1
pyudev==0.21.0
pyxdg==0.25
PyYAML==3.12
requests==2.27.1
requests-file==1.4.3
requests-ftp==0.3.1
rpm==4.14.3
selinux==2.9
sepolicy==1.1
setools==4.3.0
setroubleshoot==1.1
simpleline==1.1.2
six==1.11.0
slip==0.6.4
slip.dbus==0.6.4
sos==4.4
soupsieve==2.3.2.post1
sqlparse==0.4.2
SSSDConfig==2.7.3
syspurpose==1.28.28
systemd-python==234
tablib==3.1.0
typing_extensions==4.1.1
urllib3==1.26.12
virtualenv==20.16.7
webencodings==0.5.1
xlrd==2.0.1
xlwt==1.3.0
zipp==3.6.0
